+++ know_subnet/lm/mask.py
@@ -182,7 +182,7 @@ class MaskedHuggingfaceConv1D(Conv1D):
             masked_weight = masked_weight.reshape(self.in_features, self.out_features)
         
         size_out = x.size()[:-1] + (self.nf,)
-        x = torch.addmm(self.bias, x.view(-1, x.size(-1)), masked_weight)
+        x = torch.addmm(self.bias, x.view(-1, x.size(-1)), masked_weight.to(dtype=x.dtype))
         act = x.view(*size_out)
         return act
